createCountrySelector implementation start
